--- week3/app/src/main/java/com/mobile/week3/adapter/AlbumVPAdapter.kt
@@ -1,21 +1,0 @@
-package com.mobile.week3.adapter
-
-import androidx.fragment.app.Fragment
-import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.mobile.week3.viewpager.album.DetailFragment
-import com.mobile.week3.viewpager.album.SongFragment
-import com.mobile.week3.viewpager.album.VideoFragment
-
-class AlbumVPAdapter(fragment: Fragment) : FragmentStateAdapter(fragment) {
-
-    override fun getItemCount(): Int = 3
-
-    override fun createFragment(position: Int): Fragment {
-        return when(position) {
-            0 -> SongFragment()
-            1 -> DetailFragment()
-            else -> VideoFragment()
-        }
-    }
-
-}
