--- week3/app/src/main/java/com/mobile/week3/adapter/BannerVPAdapter.kt
@@ -1,19 +1,0 @@
-package com.mobile.week3.adapter
-
-import androidx.fragment.app.Fragment
-import androidx.viewpager2.adapter.FragmentStateAdapter
-
-class BannerVPAdapter(fragment: Fragment) : FragmentStateAdapter(fragment) {
-
-    private val fragmentList: ArrayList<Fragment> = ArrayList()
-
-    override fun getItemCount(): Int = fragmentList.size
-
-    override fun createFragment(position: Int): Fragment = fragmentList[position]
-
-    fun addFragment(fragment: Fragment) {
-        fragmentList.add(fragment)
-        notifyItemInserted(fragmentList.size - 1)   // 새로운 값이 추가된 위치 알려주는 코드
-    }
-
-}
