--- week3/app/src/main/java/com/mobile/week3/adapter/LockerVPAdapter.kt
@@ -1,20 +1,0 @@
-package com.mobile.week3.adapter
-
-import androidx.fragment.app.Fragment
-import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.mobile.week3.viewpager.locker.MusicFileFragment
-import com.mobile.week3.viewpager.locker.SavedAlbumFragment
-import com.mobile.week3.viewpager.locker.SavedSongFragment
-
-class LockerVPAdapter(fragment: Fragment) : FragmentStateAdapter(fragment) {
-
-    override fun getItemCount(): Int = 3
-
-    override fun createFragment(position: Int): Fragment {
-        return when(position) {
-            0 -> SavedSongFragment()
-            1 -> MusicFileFragment()
-            else -> SavedAlbumFragment()
-        }
-    }
-}
