--- week3/app/src/main/java/com/mobile/week3/viewpager/album/DetailFragment.kt
@@ -1,21 +1,0 @@
-package com.mobile.week3.viewpager.album
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.mobile.week3.databinding.FragmentDetailBinding
-
-class DetailFragment : Fragment() {
-
-    lateinit var binding: FragmentDetailBinding
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
-    ): View? {
-        binding = FragmentDetailBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-}
