--- week3/app/src/main/java/com/mobile/week3/viewpager/home/BannerFragment.kt
@@ -1,23 +1,0 @@
-package com.mobile.week3.viewpager.home
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.mobile.week3.databinding.FragmentBannerBinding
-
-class BannerFragment(val imgRes: Int) : Fragment() {
-
-    lateinit var binding: FragmentBannerBinding
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?,
-    ): View {
-        binding = FragmentBannerBinding.inflate(inflater, container, false)
-
-        binding.bannerImageIv.setImageResource(imgRes)
-
-        return binding.root
-    }
-}
