+++ week3/app/src/main/java/com/mobile/week4/LockerFragment.kt
@@ -1,0 +1,31 @@
+package com.mobile.week4
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.google.android.material.tabs.TabLayoutMediator
+import com.mobile.week4.adapter.LockerVPAdapter
+import com.mobile.week4.databinding.FragmentLockerBinding
+
+class LockerFragment : Fragment() {
+
+    lateinit var binding: FragmentLockerBinding
+    private val information = arrayListOf("저장한곡", "음악파일", "저장앨범")
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?,
+    ): View? {
+        binding = FragmentLockerBinding.inflate(inflater, container, false)
+
+        val lockerAdapter = LockerVPAdapter(this)
+        binding.lockerContentVp.adapter = lockerAdapter
+        TabLayoutMediator(binding.lockerContentTb, binding.lockerContentVp) {
+            tab, position ->
+            tab.text = information[position]
+        }.attach()
+
+        return binding.root
+    }
+}
