+++ week3/app/src/main/java/com/mobile/week4/MainActivity.kt
@@ -1,0 +1,84 @@
+package com.mobile.week4
+
+import android.content.Intent
+import android.os.Bundle
+import android.view.View
+import android.widget.Toast
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import com.mobile.week4.data.AlbumDto
+import com.mobile.week4.databinding.ActivityMainBinding
+
+class MainActivity : AppCompatActivity() {
+    lateinit var binding: ActivityMainBinding
+    val SONG_CODE = 100
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        // Fragment 초기화
+        supportFragmentManager.beginTransaction()
+            .replace(R.id.main_frm, HomeFragment())
+            .commitAllowingStateLoss()
+
+        // Fragment 전환
+        binding.mainBnv.setOnItemSelectedListener { item ->
+            when (item.itemId) {
+                R.id.homeFragment -> {
+                    supportFragmentManager.beginTransaction()
+                        .replace(R.id.main_frm, HomeFragment())
+                        .commitAllowingStateLoss()
+                    return@setOnItemSelectedListener true
+                }
+                R.id.lockerFragment -> {
+                    supportFragmentManager.beginTransaction()
+                        .replace(R.id.main_frm, LockerFragment())
+                        .commitAllowingStateLoss()
+                    return@setOnItemSelectedListener true
+                }
+            }
+            false
+        }
+
+        // Activity 전환
+        binding.mainPlayerCl.setOnClickListener {
+            val intent = Intent(this, SongActivity::class.java)
+            intent.putExtra("songId", 1)
+            startActivityForResult(intent, SONG_CODE)
+        }
+    }
+
+    override fun onActivityResult( requestCode: Int, resultCode: Int, data: Intent? ) {
+        super.onActivityResult(requestCode, resultCode, data)
+
+        if (requestCode == SONG_CODE) {
+            when (resultCode) {
+                RESULT_OK -> {
+                    val result = data?.getStringExtra("result_song")
+                    Toast.makeText(this, "Song: $result", Toast.LENGTH_SHORT).show()
+                }
+                else -> {
+                    Toast.makeText(this, "resultCode가 올바르지 않습니다.", Toast.LENGTH_SHORT).show()
+                }
+            }
+        }
+    }
+
+    fun updateMiniPlayer(album: AlbumDto) {
+        binding.mainMiniplayerTitleTv.text = album.title
+        binding.mainMiniplayerSingerTv.text = album.singer
+
+        if (binding.mainMiniplayerBtn.visibility == View.GONE) {
+            binding.mainMiniplayerBtn.visibility = View.VISIBLE
+            binding.mainPauseBtn.visibility = View.GONE
+        }
+        else {
+            binding.mainMiniplayerBtn.visibility = View.GONE
+            binding.mainPauseBtn.visibility = View.VISIBLE
+        }
+
+    }
+}
