+++ week3/app/src/main/java/com/mobile/week4/adapter/AlbumLockerRVAdapter.kt
@@ -1,0 +1,63 @@
+package com.mobile.week4.adapter
+
+import android.annotation.SuppressLint
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import com.mobile.week4.data.AlbumDto
+import com.mobile.week4.databinding.ItemLockerAlbumBinding
+
+class AlbumLockerRVAdapter(): RecyclerView.Adapter<AlbumLockerRVAdapter.ViewHolder>() {
+    private val albums = ArrayList<AlbumDto>()
+
+    interface MyItemClickListener{
+        fun onRemoveSong(songId: Int)
+    }
+
+    private lateinit var mItemClickListener: MyItemClickListener
+
+    fun setMyItemClickListener(itemClickListener: MyItemClickListener){
+        mItemClickListener = itemClickListener
+    }
+
+    override fun onCreateViewHolder(viewGroup: ViewGroup, viewType: Int): AlbumLockerRVAdapter.ViewHolder {
+        val binding: ItemLockerAlbumBinding = ItemLockerAlbumBinding.inflate(
+            LayoutInflater.from(viewGroup.context),
+            viewGroup, false
+        )
+
+        return ViewHolder(binding)
+    }
+
+    override fun onBindViewHolder(holder: AlbumLockerRVAdapter.ViewHolder, position: Int) {
+        holder.bind(albums[position])
+        holder.binding.itemAlbumMoreIv.setOnClickListener {
+            mItemClickListener.onRemoveSong(albums[position].id)
+            removeSong(position)
+        }
+    }
+
+    override fun getItemCount(): Int = albums.size
+
+    @SuppressLint("NotifyDataSetChanged")
+    fun addAlbums(albums: ArrayList<AlbumDto>) {
+        this.albums.clear()
+        this.albums.addAll(albums)
+
+        notifyDataSetChanged()
+    }
+
+    fun removeSong(position: Int){
+        albums.removeAt(position)
+        notifyDataSetChanged()
+    }
+
+    inner class ViewHolder(val binding: ItemLockerAlbumBinding) : RecyclerView.ViewHolder(binding.root){
+        fun bind(album: AlbumDto){
+            binding.itemAlbumImgIv.setImageResource(album.coverImg!!)
+            binding.itemAlbumTitleTv.text = album.title
+            binding.itemAlbumSingerTv.text = album.singer
+        }
+    }
+
+}
