+++ week3/app/src/main/java/com/mobile/week4/data/AlbumDto.kt
@@ -1,0 +1,12 @@
+package com.mobile.week4.data
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity("AlbumTable")
+data class AlbumDto(
+    @PrimaryKey(autoGenerate = false) var id: Int = 0,
+    var title: String? = "제목",
+    var singer: String? = "가수",
+    var coverImg: Int? = null
+)
