+++ week3/app/src/main/java/com/mobile/week4/view_model/SharedViewModel.kt
@@ -1,0 +1,15 @@
+package com.mobile.week4.view_model
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import com.mobile.week4.data.AlbumDto
+
+class SharedViewModel : ViewModel() {
+    private val _selectedAlbum = MutableLiveData<AlbumDto>()
+    val selectedAlbum: LiveData<AlbumDto> get() = _selectedAlbum
+
+    fun setAlbum(value: AlbumDto) {
+        _selectedAlbum.value = value
+    }
+}
