+++ week3/app/src/main/java/com/mobile/week4/viewpager/album/SongFragment.kt
@@ -1,0 +1,31 @@
+package com.mobile.week4.viewpager.album
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.mobile.week4.databinding.FragmentSongBinding
+
+class SongFragment : Fragment() {
+
+    lateinit var binding: FragmentSongBinding
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
+    ): View? {
+        binding = FragmentSongBinding.inflate(inflater, container, false)
+
+        binding.songMixoffTg.setOnClickListener {
+            binding.songMixonTg.visibility = View.VISIBLE
+            binding.songMixoffTg.visibility = View.GONE
+        }
+
+        binding.songMixonTg.setOnClickListener {
+            binding.songMixoffTg.visibility = View.VISIBLE
+            binding.songMixonTg.visibility = View.GONE
+        }
+
+        return binding.root
+    }
+}
