+++ week3/app/src/main/java/com/mobile/week4/viewpager/album/VideoFragment.kt
@@ -1,0 +1,20 @@
+package com.mobile.week4.viewpager.album
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.mobile.week4.databinding.FragmentVideoBinding
+
+class VideoFragment : Fragment() {
+
+    lateinit var binding: FragmentVideoBinding
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
+    ): View? {
+        binding = FragmentVideoBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+}
